+++ src/sat/ipasir.zig
@@ -226,7 +226,7 @@ pub const IpasirSolver = struct {
     }
 
     pub fn signature() [*:0]const u8 {
-        return "logic-zig-ipasir-0.12";
+        return "logic-zig-ipasir-1.0";
     }
 };